ditch PrayerContent component and move input logic to PrayerInput component
