+++ packages/apollos-ui-prayer/src/PrayerCard/index.js
@@ -14,7 +14,7 @@ import {
   withTheme,
 } from '@apollosproject/ui-kit';
 
-// import PrayerContent from './PrayerContent';
+import PrayerInput from './PrayerInput';
 
 const AvatarWrapper = styled(
   ({ hasAvatar, theme }) => ({
@@ -51,22 +51,6 @@ const Content = styled(({ theme }) => ({
   paddingVertical: theme.sizing.baseUnit * 1.5, // TODO: move this style into `CardContent`
 }))(CardContent);
 
-const PrayerPrompt = styled({
-  alignItems: 'center',
-  flexDirection: 'row',
-})(View);
-
-const PlusIcon = withTheme(
-  ({ theme }) => ({
-    name: 'plus',
-    size: theme.sizing.baseUnit * 1.5,
-    style: {
-      marginRight: theme.sizing.baseUnit * 0.5,
-    },
-  }),
-  'ui-prayer.PrayerCard.PlusIcon'
-)(Icon);
-
 const PrayerCard = ({ avatar, prayer, title }) => (
   <StyledCard>
     <Content>
@@ -79,15 +63,7 @@ const PrayerCard = ({ avatar, prayer, title }) => (
       </AvatarWrapper>
 
       <H4 padded>{title}</H4>
-      {prayer ? (
-        <BodyText>{prayer}</BodyText>
-      ) : (
-        <PrayerPrompt>
-          <PlusIcon />
-          <BodyText>{`I'm thankful for...`}</BodyText>
-        </PrayerPrompt>
-      )}
-      {/* <PrayerContent title={title} prayer={prayer} /> */}
+      {prayer ? <BodyText>{prayer}</BodyText> : <PrayerInput />}
     </Content>
   </StyledCard>
 );
